feedback change rate from -1 and 1 to between 0 and 1 try to get feedback from another AI api for learning

+++ app/db/schema.sql
@@ -38,7 +38,7 @@ CREATE TABLE IF NOT EXISTS feedback (
     id INTEGER PRIMARY KEY AUTOINCREMENT,
     source_book_id INTEGER NOT NULL,
     candidate_book_id INTEGER NOT NULL,
-    label INTEGER NOT NULL,
+    label FLOAT NOT NULL,
     created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
     UNIQUE(source_book_id, candidate_book_id),
     FOREIGN KEY (source_book_id) REFERENCES books(id),
